V5.5-000

+++ sr_unix/check_encrypt_support.sh
@@ -26,6 +26,7 @@ if [ "OSF1" = "$hostos" -o \( "HP-UX" = "$hostos" -a "ia64" != `uname -m` \) ]; 
 fi
 
 lib_search_path="/usr/local/lib64 /usr/local/lib /usr/lib64 /usr/lib /lib64 /lib /usr/local/ssl/lib /usr/lib/x86_64-linux-gnu"
+lib_search_path="$lib_search_path /lib/x86_64-linux-gnu"
 include_search_path="/usr/include /usr/local/include /usr/local/include/gpgme"
 bin_search_path="/usr/bin /usr/local/bin /bin"
 
@@ -101,41 +102,55 @@ do
 	fi
 done
 
-hostname=`uname -n | awk -F. '{print $1}'`
-if [ $found_headers = "TRUE" -a $found_libs = "TRUE" -a $found_bins = "TRUE" ];then
-#	We want to keep at least one server without support for encryption.  So if it send a warning if it is not the case.
-	non_encrypt_other_servers="dincern"
-	echo $non_encrypt_other_servers | grep -w $hostname > /dev/null
-	if [ $? -eq 0 ]; then
-		errmsg="This system should not supports encryption, but it does.\n"
-		echo $errmsg | mailx -s \
-			"ENCRYPTSUPPORTED-W-WARNING : Server $hostname will build encryption plugin (but should not)" \
-		 	gglogs
+server_list="$cms_tools/server_list"
+
+if [ -f $server_list ];then
+	hostname=`uname -n | awk -F. '{print $1}'`
+	eval `/usr/local/bin/tcsh -efc "
+		source $server_list;
+		echo non_encrypt_machines=\'\\\$non_encrypt_machines\';
+		echo encrypt_desktops=\'\\\$encrypt_desktops\';
+		echo encrypt_dist_servers=\'\\\$encrypt_dist_servers\';
+		echo encrypt_other_servers=\'\\\$encrypt_other_servers\';
+	" || echo echo ERROR \; exit 1`
+	if [ $found_headers = "TRUE" -a $found_libs = "TRUE" -a $found_bins = "TRUE" ];then
+	#	We want to keep at least one server without support for encryption.  So if it send a warning if it is not the case.
+		echo $non_encrypt_machines | grep -w $hostname > /dev/null
+		if [ $? -eq 0 ]; then
+			errmsg="This system should not supports encryption, but it does.\n"
+			printf "$errmsg" | mailx -s \
+				"ENCRYPTSUPPORTED-W-WARNING : Server $hostname will build encryption plugin (but should not)" \
+				gglogs
+		fi
+		echo "TRUE"
+	else
+	# 	These are distribution servers that support encryption. If can't build encryption plugin send error.
+		echo $encrypt_dist_servers | grep -w $hostname > /dev/null
+		if [ $? -eq 0 ]; then
+			errmsg="Please setup the required dependencies for this distribution server:\n"
+			errmsg="$errmsg$whatsmissing"
+			printf "$errmsg" | mailx -s \
+				"ENCRYPTSUPPORTED-E-ERROR : Distribution server $hostname will not build encryption plugin" \
+				gglogs
+		fi
+	#	There are machines that can support encryption. If can't build encryption plugin send warning.
+		encrypt_machines="$encrypt_other_servers $encrypt_desktops"
+		echo $encrypt_machines | grep -w $hostname > /dev/null
+		if [ $? -eq 0 ]; then
+			errmsg="This system supports encryption but does not have the required dependencies setup:\n"
+			errmsg="$errmsg$whatsmissing"
+			printf "$errmsg" | mailx -s \
+				"ENCRYPTSUPPORTED-W-WARNING : Server $hostname will not build encryption plugin" \
+				gglogs
+		fi
+		echo "FALSE"
 	fi
-	echo "TRUE"
 else
-# 	These are distribution servers that support encryption. If can't build encryption plugin send error.
-	encrypt_dist_servers="charybdis jackal pfloyd snail atllita1 atlhxit1 sagaloo"
-	echo $encrypt_dist_servers | grep -w $hostname > /dev/null
-	if [ $? -eq 0 ]; then
-		errmsg="Please setup the required dependencies for this distribution server:\n"
-		errmsg="$errmsg $whatsmissing"
-		echo $errmsg | mailx -s \
-			"ENCRYPTSUPPORTED-E-ERROR : Distribution server $hostname will not build encryption plugin" \
-			gglogs
+	if [ $found_headers = "TRUE" -a $found_libs = "TRUE" -a $found_bins = "TRUE" ];then
+		echo "TRUE"
+	else
+		echo "FALSE"
 	fi
-#	There are servers that can support encryption. If can't build encryption plugin send warning.
-	encrypt_other_servers="scylla mlnlit1 lespaul atlst2000 turtle atllita2 atlhxit2 sagapaneer"
-	encrypt_other_servers="$encrypt_other_servers rajamanin kishoreh johnsons shaha maimoneb estess sopini karthikk bahirs"
-	encrypt_other_servers="$encrypt_other_servers parenteaul"
-	echo $encrypt_other_servers | grep -w $hostname > /dev/null
-	if [ $? -eq 0 ]; then
-		errmsg="This system supports encryption but does not have the required dependencies setup:\n"
-		errmsg="$errmsg $whatsmissing"
-		echo $errmsg | mailx -s \
-			"ENCRYPTSUPPORTED-W-WARNING : Server $hostname will not build encryption plugin" \
-		 	gglogs
-	fi
-	echo "FALSE"
 fi
+
 exit 0
